requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
streamlit
tensorflow==2.19.0
numpy
pillow
kaggle==1.7.4.5